extract menu into a separat component
